Eliminated bug on float accumulator, replaced by a double accumulator.

+++ Saxs/saxsKernel.h
@@ -48,7 +48,7 @@ public:
     void scaledCell();
     void zeroIq();
     void getHistogram(std::vector<std::vector<float>> &);
-    std::vector<std::vector<float>> getSaxs();
+    std::vector<std::vector<double>> getSaxs();
     void createMemory();
     void resetHistogramParameters(std::vector<std::vector<float>> &);
     void writeBanner();
@@ -85,13 +85,13 @@ private:
     thrust::device_vector<cuFloatComplex> d_Iq;
     thrust::device_vector<cuFloatComplex> d_gridSupC;
     thrust::device_vector<float> d_histogram;
-    thrust::device_vector<float> d_nhist;
+    thrust::device_vector<double> d_nhist;
 
     thrust::host_vector<float> h_moduleX;
     thrust::host_vector<float> h_moduleY;
     thrust::host_vector<float> h_moduleZ;
     thrust::host_vector<float> h_histogram;
-    thrust::host_vector<float> h_nhist;
+    thrust::host_vector<double> h_nhist;
 
     float *d_grid_ptr{nullptr};
     float *d_gridSup_ptr{nullptr};
@@ -103,7 +103,7 @@ private:
     float *d_moduleY_ptr{nullptr};
     float *d_moduleZ_ptr{nullptr};
     float *d_histogram_ptr{nullptr};
-    float *d_nhist_ptr{nullptr};
+    double *d_nhist_ptr{nullptr};
     std::function<int(int, float)> borderBins = [](int nx, float shell) -> int
     {
         return static_cast<int>(shell * nx / 2);
